1.3.1-alpha.2
 - constants file
 - releases docs
 - split out tests
