+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "apialerts",
-  "version": "1.3.1",
+  "version": "1.3.1-alpha.2",
   "description": "JavaScript/TypeScript SDK for apialerts.com",
   "author": "API Alerts",
   "license": "MIT",
