+++ src/constants.ts
@@ -1,0 +1,3 @@
+export const INTEGRATION = 'js'
+export const VERSION = '1.3.1'
+export const BASE_URL = 'https://api.apialerts.com/event'
